requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1

gspread==6.2.1
google-auth==2.48.0

playwright==1.58.0
requests==2.32.5
